修复缓存复用与生图时的维度问题
修复点：

将 Comfy 节点加载 Lance 时的 latent_patch_size 固定为官方 demo 实际使用的 [1, 1, 1]，和本地 checkpoint 的 vae2llm.weight [2048, 48]、llm2vae.weight [48, 2048] 对齐，解决 mat1 and mat2 shapes cannot be multiplied (1024x192 and 48x2048)。
量化缓存加载前新增 shape 预检；如果缓存和当前模型结构不匹配，会自动放弃缓存并重建，不会再出现“元数据通过但 load_state_dict 炸掉后要求手动重建”的情况。
分块量化 checkpoint 时也加了 shape 检查，避免错误结构被量化成一个后面推理才爆的缓存。
3D 位置编码中间数组从 float64 改为 float32，降低 CPU 内存峰值，缓解你日志里的 NumPy allocation error。

+++ Lance/modeling/lance/modeling_utils.py
@@ -56,11 +56,11 @@ def get_1d_sincos_pos_embed_from_grid(embed_dim, pos):
     out: (M, D)
     """
     assert embed_dim % 2 == 0
-    omega = np.arange(embed_dim // 2, dtype=np.float64)
+    omega = np.arange(embed_dim // 2, dtype=np.float32)
     omega /= embed_dim / 2.
     omega = 1. / 10000**omega  # (D/2,)
 
-    pos = pos.reshape(-1)  # (M,)
+    pos = pos.reshape(-1).astype(np.float32, copy=False)  # (M,)
     out = np.einsum('m,d->md', pos, omega)  # (M, D/2), outer product
 
     emb_sin = np.sin(out) # (M, D/2)
